o) Add a UFS_BSIZE macro which does 1<<bshift. o) Add UFS on-disk directory structure. o) Add a context for iterating through a directory. o) Track the current inode. o) Add functions to read data from inodes and entries from directories. o) List all items in the root directory on startup. o) Use uint32_t consistently for inodes.

+++ mk/io/storage/ufs/ufs_inode.c
@@ -10,7 +10,7 @@ COMPILE_TIME_ASSERT(offsetof(struct ufs2_inode, in_indirect) == 0xd0);
 COMPILE_TIME_ASSERT(sizeof (struct ufs2_inode) == 256);
 
 uint64_t
-ufs_inode_block(struct ufs_superblock *sb, uint64_t inode)
+ufs_inode_block(struct ufs_superblock *sb, uint32_t inode)
 {
 	uint64_t cg = inode / sb->sb_ipg;
 	uint64_t ig = (cg * sb->sb_fpg) + sb->sb_iboff;
